add get reviews test, install supertest, separated port connection from app routes for testing
