+++ package.json
@@ -4,7 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "jest",
+    "coverage": "jest --coverage",
+    "test": "jest --verbose",
     "start": "nodemon server/index.js"
   },
   "repository": {
@@ -23,6 +24,7 @@
     "pg": "^8.6.0"
   },
   "devDependencies": {
-    "jest": "^27.0.6"
+    "jest": "^27.0.6",
+    "supertest": "^6.1.4"
   }
 }
